refactor(shared/db): reorganiza modelos en subcarpetas cv/visitor/taxonomy/i18n
- Crea shared/db/models/{cv,visitor,taxonomy,i18n}/ con un archivo por
  agrupacion logica
- cv/: profile, experience, education, project, skill, cv_entity
  (awards/certs/langs/pubs/refs + sus junctions)
- visitor/: contact, session, session_visit, tracking
- taxonomy/: catalog (niches, tech_tags), priority (niche_priorities),
  event_type
- i18n/: translation
- Junctions de cada entidad CV se mueven al archivo de su entidad padre
  (no mas junctions.py monolitico)
- Elimina los 11 archivos viejos en shared/db/models/*.py
- __init__.py raiz re-exporta todo para preservar API publica:
  from shared.db.models import Profile, Contact, Niche, etc. sigue
  funcionando sin cambio en call-sites
- __tablename__ sin cambios (DB intacta, solo refactor Python)
- Base.metadata sigue con las 37 tablas
