+++ serverless/lambda/shared/db/models/cv/profile.py
@@ -1,4 +1,4 @@
-"""@module profile — singleton de perfil + sus stats + sus niches.
+"""@module cv.profile — singleton de perfil + stats + junction con niches.
 
 `profile` es un singleton (siempre 1 fila). Sus textos bilingues
 (`headline`, `summary`, `availability`) NO viven aqui: van a `translations`.
@@ -9,7 +9,7 @@ del profile se persisten en la union `profile_niches`.
 from sqlalchemy import ForeignKey, Integer, PrimaryKeyConstraint, String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from ..base import Base, TimestampMixin, UUIDPKMixin
+from ...base import Base, TimestampMixin, UUIDPKMixin
 
 
 class Profile(UUIDPKMixin, TimestampMixin, Base):
@@ -23,9 +23,7 @@ class Profile(UUIDPKMixin, TimestampMixin, Base):
 
     name: Mapped[str] = mapped_column(String(160), nullable=False)
     handle: Mapped[str] = mapped_column(String(80), nullable=False, unique=True)
-    # Ciudad y pais en formato limpio (ej. "Lima, Perú").
     location: Mapped[str] = mapped_column(String(160), nullable=False)
-    # Contactos.
     email: Mapped[str] = mapped_column(String(254), nullable=False)
     phone: Mapped[str | None] = mapped_column(String(40))
     linkedin_url: Mapped[str] = mapped_column(String(500), nullable=False)
@@ -36,7 +34,7 @@ class Profile(UUIDPKMixin, TimestampMixin, Base):
 
 class ProfileStats(UUIDPKMixin, TimestampMixin, Base):
     """Stats declarados del perfil (cards de la StatsBar). Relacion 1:1
-    con `profile` — se separa porque es un grupo cohesivo opcional.
+    con `profile`.
     """
 
     __tablename__ = 'profile_stats'
@@ -53,12 +51,7 @@ class ProfileStats(UUIDPKMixin, TimestampMixin, Base):
 
 
 class ProfileNiche(Base):
-    """Union profile <-> niche. El profile es singleton, pero sus niches
-    se persisten igual para que la DB sea fuente de verdad completa del CV.
-
-    Union pura `(profile_id, niche_id)` con PK compuesta — mismo patron
-    que las demas tablas `<entidad>_niches` de `junctions.py`.
-    """
+    """Union profile <-> niche."""
 
     __tablename__ = 'profile_niches'
