+++ serverless/lambda/shared/db/models/i18n/__init__.py
@@ -1,0 +1,5 @@
+"""Re-exports planos del dominio i18n."""
+
+from .translation import Translation
+
+__all__ = ['Translation']
