+++ serverless/lambda/shared/db/models/taxonomy/__init__.py
@@ -1,0 +1,7 @@
+"""Re-exports planos del dominio taxonomy."""
+
+from .catalog import Niche, TechTag
+from .event_type import EventType
+from .priority import NichePriority
+
+__all__ = ['EventType', 'Niche', 'NichePriority', 'TechTag']
