+++ serverless/lambda/shared/db/models/visitor/__init__.py
@@ -1,0 +1,8 @@
+"""Re-exports planos del dominio visitor."""
+
+from .contact import Contact
+from .session import Session
+from .session_visit import SessionVisit
+from .tracking import TrackingEvent
+
+__all__ = ['Contact', 'Session', 'SessionVisit', 'TrackingEvent']
